+++ CHANGELOG.md
@@ -1,5 +1,23 @@
 # Arbimon Release Notes
 
+## v3.0.35 - August XX, 2021
+
+Resolved issues:
+
+- CE-1183 Fixed filter in the recordings page. The user is able to select the month of January
+
+## v3.0.34 - August XX, 2021
+
+New features:
+
+- CE-1101 A search bar is updated for projects across Arbimon
+- CE-1136 A Summary stats data is added to the Home page
+- CE-1105 The project list is updated on the Home page
+
+Resolved issues:
+
+- CE-1197 Change wording when choosing filename format in the uploads page
+
 ## v3.0.33-hotfix.0 - August 03, 2021
 
 Resolved issues: